# Improved performance of boolean comparison (5x-14x)