The idea is that we can compare full bytes instead of iterating over individual bits, replacing `Fn(bool,bool) -> bool` by `Fn(u8,u8) -> u8` with corresponding bit operations (e.g. `a != b -> a ^ b`, `a < b -> !a & b`).

```
git checkout 0893100640f22dfc5fec4dede300f63fe095b170
cargo bench --no-default-features --features benchmarks,compute --bench comparison_kernels -- "bool.*2\^20"
git checkout cmp_boolean
cargo bench --no-default-features --features benchmarks,compute --bench comparison_kernels -- "bool.*2\^20"
```

```
bool 2^20               time:   [1.0651 ms 1.0719 ms 1.0817 ms]                       
                        change: [-78.884% -78.378% -77.783%] (p = 0.00 < 0.05)

bool scalar 2^20        time:   [212.37 us 213.76 us 215.72 us]                             
                        change: [-93.123% -93.040% -92.965%] (p = 0.00 < 0.05)
```